Assistant checkpoint: Update schema and form for incoming loads
Assistant generated file changes:
- shared/schema.ts: Update loads table schema for incoming loads
- client/src/components/loads/IncomingLoadForm.tsx: Update incoming load form

---

User prompt:

That is wrong, in my prior change i said we do not need Deliverylocation and pickup location anymore. Please recode the entire database to only include these columns for INCOMING loads. The Incoming, wholesale, and miscellaneous forms are independent and should be stored on different databases. They should not be conflicting with eachother.

For incoming load type (we will work on the other 2 later) we only need to be able to add the following fields.

- Load Type Status can be removed (we will have a status toggle the load has already been added)
- At the top we need to add a "Supplier" field
- Pickuplocation and Deliverylocation can be merged into 1. Lets call this "Location". This should have 2 options, British Columbia or Ontario because we have 2 warehouses.
- We dont need Driver Name and Phone
- Instead of "equipment", call it "notes"
- Add "Load Cost" right above freight cost, In the database, this should create another column called total costs that sums both of them up
- We also need another field called "Profit ROI"
- We dont need container number, booking number, vessel name, voyage number or estiamted port of arrival
- We need to be able to add the following file attachments: BOL, Material Invoice, Freight Invoice, and Load Performance.

This is only in regards to the "Incoming" Load type, we will update the others later on.